+++ package.json
@@ -2,27 +2,24 @@
   "name": "pldb",
   "description": "PLDB: a Programming Language Database",
   "scripts": {
-    "tsc": "cd code; tsc; cd ..; npm run plusX",
+    "tsc": "cd code; tsc; cd ..; npm run makeCrawlersExecutable",
     "test": "./code/test.js",
-    "build": "chmod +x ./code/SiteBuilder.js; ./code/SiteBuilder.js buildAll",
-    "open": "open site/index.html",
     "search": "./code/cliSearch.js",
     "local": "./code/PLDBServer.js startDevServer",
-    "format": "./code/SiteBuilder.js formatDatabase",
+    "format": "./code/PLDBServer.js formatDatabase",
     "crawl": "./code/crawlers/Crawler.js crawl",
     "githubs": "./code/crawlers/github.com/GitHub.js fetchAllRepoData && ./code/crawlers/github.com/GitHub.js writeAllRepoData",
     "fetchContributors": "wget -O site/pages/contributors.json https://api.github.com/repos/breck7/pldb/contributors",
-    "buildCombined": "cat site/frontEndJavascript/mousetrap.js > site/frontEndJavascript/combined.js; echo '\n\n' >> site/frontEndJavascript/combined.js; cat site/frontEndJavascript/autocomplete.js >> site/frontEndJavascript/combined.js; echo '\n\n' >> site/frontEndJavascript/combined.js; cat site/frontEndJavascript/search.js >> site/frontEndJavascript/combined.js",
-    "perfBuild": "time node --cpu-prof --cpu-prof-name=ignore/pldbBuildNodePerf.cpuprofile ./code/SiteBuilder.js buildAll",
     "clean": "cd code; tsc --build --clean",
     "cloc": "cloc --vcs git . --read-lang-def=code/clocLangs.txt",
     "checkLinks": "linkinator -r https://pldb.com > ignore/brokenLinks.txt",
-    "deploy": "echo 'Rsyncing site/ to /var/www/html on Marconi'; rsync -vrt --progress site/* pldb.com:/var/www/html/",
     "jtreeBeta": "npm install ~/jtree",
     "scrollBeta": "npm install ~/scroll",
-    "fetchBuildLogs": "scp file pldb.com:/home/pldb/pldb/ignore/treeBaseServerLog.tree ignore/",
-    "plusX": "chmod u+x code/*.js; chmod u+x code/**/*.js; chmod u+x code/**/*.js; chmod u+x code/**/**/*.js; chmod u-x code/treeBaseFrontEndApp.js",
+    "makeCrawlersExecutable": "chmod u+x code/**/*.js; chmod u+x code/**/*.js; chmod u+x code/**/**/*.js",
     "publish": "cd code; npm publish",
+    "deploy": "ssh pldb.com; git pull; x tsc; x restart",
+    "fetchLogs": "scp file pldb.com:/root/pldb/ignore/treeBaseServerLog.tree ignore/prodLogs/",
+    "prod": "pm2 start ./code/PLDBServer.js -- startProdServer",
     "restart": "pm2 restart 0"
   },
   "prettier": {
