+++ site/frontEndJavascript/readme.scroll
@@ -2,11 +2,6 @@
 
 * This folder contains 3rd party JS libs used by our site and a few small scripts that glue them to enhance the basic functionality.
 
-# Building the combined file:
-
-bashCode
- npm run buildCombined
-
 # Libraries
 
 ## Autocomplete by kraaden
